front-end/fb-functions/functions/requirements.txt

google-generativeai>=0.3.0
google-api-core>=2.11.0
google-cloud-secret-manager>=2.16.0
functions-framework
firebase_functions~=0.1.0
pydantic>=2.0.0
